requirements/app.txt
guniflask-cli>=0.9.2,<0.10
Flask-SQLAlchemy>=2.4.1
PyMySQL>=0.9.3
pyyaml>=5.3.1
